refactor(permissions): Move types to config
This is to make it easier to discover corresponding permission functions.
