+++ apps/web/src/components/tx-flow/flows/TokenTransfer/__tests__/CreateTokenTransfer.test.tsx
@@ -2,7 +2,7 @@ import { TokenTransferType } from '@/components/tx-flow/flows/TokenTransfer'
 import { CreateTokenTransfer } from '@/components/tx-flow/flows/TokenTransfer/CreateTokenTransfer'
 import * as tokenUtils from '@/components/tx-flow/flows/TokenTransfer/utils'
 import * as useHasPermission from '@/permissions/hooks/useHasPermission'
-import { Permission } from '@/permissions/types'
+import { Permission } from '@/permissions/config'
 import { render } from '@/tests/test-utils'
 import { ZERO_ADDRESS } from '@safe-global/protocol-kit/dist/src/utils/constants'
