--- apps/web/src/permissions/types/Permission.ts
@@ -1,8 +1,0 @@
-export enum Permission {
-  CreateTransaction = 'CreateTransaction',
-  ProposeTransaction = 'ProposeTransaction',
-  SignTransaction = 'SignTransaction',
-  ExecuteTransaction = 'ExecuteTransaction',
-  CreateSpendingLimitTransaction = 'CreateSpendingLimitTransaction',
-  EnablePushNotifications = 'EnablePushNotifications',
-}
