--- apps/web/src/permissions/types/Role.ts
@@ -1,10 +1,0 @@
-export enum Role {
-  Owner = 'Owner',
-  NestedOwner = 'NestedOwner',
-  Proposer = 'Proposer',
-  Executioner = 'Executioner',
-  ModuleRole = 'ModuleRole',
-  Recoverer = 'Recoverer',
-  SpendingLimitBeneficiary = 'SpendingLimitBeneficiary',
-  NoWalletConnected = 'NoWalletConnected',
-}
